Test and document useForm
